requirements.txt
pydantic>=2.12.3
pydantic-settings>=2.11.0
google-genai>=1.49.0
fastapi>=0.115.0
fastapi-healthchecks>=1.1.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0